% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_image.R
\name{detrend_image}
\alias{detrend_image}
\title{A Light De-Trending Function}
\usage{
detrend_image(im, outfile = "", channel = "Color", k = 1)
}
\arguments{
\item{im}{The image to adjust}

\item{outfile}{Where to put the result}

\item{channel}{Grey or Color. Defaults to Color}

\item{k}{Scale factor. Defaults to 1.0}
}
\description{
This function adjusts the light in an image by removing a linear trend.
}
