% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_band.R
\name{highlight_band}
\alias{highlight_band}
\title{A Pixel Highlighting Function}
\usage{
highlight_band(im, lo, hi)
}
\arguments{
\item{im}{The image}

\item{lo}{One threshold}

\item{hi}{Another threshold}
}
\description{
This function highlights pixels having values between two thresholds or outside of two thresholds, depending on whether the first threshold is above or below the second.
}
