% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im2cim.R
\name{im2cim}
\alias{im2cim}
\title{An Image Format Conversion Function}
\usage{
im2cim(im)
}
\arguments{
\item{im}{The image to convert}
}
\description{
This function converts image data obtained via OpenImageR::readImage() into C-image format.
}
