% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2english.R
\name{make_a_plot}
\alias{make_a_plot}
\title{A Collection of ggplot Shortcuts}
\usage{
make_a_plot(df)
}
\arguments{
\item{im}{An image}
}
\description{
These functions put commonly used ggplot commands into plain language
}
