% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2english.R
\name{plot_cimg}
\alias{plot_cimg}
\title{A C-Image Plotting Function.}
\usage{
plot_cimg(im, palette = c("black", "white"))
}
\arguments{
\item{im}{An image}

\item{palette}{A collection of colors}
}
\description{
This function plots an image (in C-image format) using a variety of color settings
}
